+++ lending/src/sdk/models/operations/createaccount.ts
@@ -3,12 +3,12 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class CreateAccountRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
-    accountingAccount?: shared.AccountingAccount;
+    accountPrototype?: shared.AccountPrototype;
 
     /**
      * Allow a sync upon push completion.