+++ lending/src/sdk/models/operations/getaccountingagedcreditorsreport.ts
@@ -3,8 +3,8 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { RFCDate } from "../../types";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
+import { RFCDate } from "../../../sdk/types";
 import { AxiosResponse } from "axios";
 
 export class GetAccountingAgedCreditorsReportRequest extends SpeakeasyBase {