+++ lending/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v3.4.0] lending
 ### Releases
 - [NPM v3.4.0] https://www.npmjs.com/package/@codat/lending/v/3.4.0 - lending
+
+## 2023-11-13 01:28:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0] lending
+### Releases
+- [NPM v4.0.0] https://www.npmjs.com/package/@codat/lending/v/4.0.0 - lending